Fixed potential edge case not reporting GameCenter auth failure
According to the documentation for [GKLocalPlayer authenticateHandler]
(https://developer.apple.com/library/ios/documentation/GameKit/Reference
/GKLocalPlayer_Ref/#//apple_ref/occ/instp/GKLocalPlayer/authenticateHand
ler), the process is to check if the user is authenticated, then check
if view controller is non-nil, and anything else is a failure. This
implies that it’s possible for a failure to be caused and have “error”
be nil.

Removed nil check and add default error message.